requirements.txt
aenum==1.4.7
agate==1.5.5
agate-dbf==0.2.0
agate-excel==0.2.0
agate-sql==0.5.0
anyjson==0.3.3
appdirs==1.4.2
awesome-slugify==1.6.5
Babel==2.3.4
billiard==3.5.0.2
boto==2.46.1
boto3==1.4.4
botocore==1.5.18
cffi==1.9.1
colorthief==0.2.1
coreapi==2.2.4
coreschema==0.0.4
coverage==4.3.4
cryptography==1.7.2
csvkit==1.0.1
cycler==0.10.0
Cython==0.25.2
dbf==0.96.8
dbfread==2.0.7
Django==1.10.5
django-appconf==1.0.2
django-compressor==2.0
django-configurations==2.0
django-crontab==0.7.1
django-environ==0.4.1
django-post-office==2.0.8
django-pyodbc-azure==1.10.4.0
django-rest-framework-mongoengine==3.3.1
django-rest-swagger==2.1.1
djangorestframework==3.5.4
docutils==0.13.1
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
gcloud==0.18.3
googleapis-common-protos==1.5.2
gunicorn==19.6.0
httplib2==0.10.3
idna==2.2
ipaddress==1.0.18
isodate==0.5.4
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.5
jmespath==0.9.1
jsonfield==1.0.3
jws==0.1.3
leather==0.3.3
MarkupSafe==0.23
matplotlib==1.5.3
Moments==2.0
mysqlclient==1.3.10
numpy==1.11.2
oauth2client==4.0.0
olefile==0.44
openapi-codec==1.3.1
openpyxl==2.4.4
packaging==16.8
parsedatetime==2.2
Pillow==4.0.0
protobuf==3.2.0
psycopg2==2.6.2
pyasn1==0.2.2
pyasn1-modules==0.0.8
pycparser==2.17
pycryptodome==3.4.5
pymongo==3.4.0
pyodbc==4.0.12
pyparsing==2.1.10
Pyrebase==3.0.27
python-dateutil==2.6.0
python-jwt==2.0.1
pytimeparse==1.1.6
pytz==2016.10
PyYAML==3.12
rcssmin==1.0.6
regex==2017.2.8
requests==2.13.0
requests-toolbelt==0.7.1
rjsmin==1.0.12
rsa==3.4.2
s3transfer==0.1.10
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.1.5
Unidecode==0.4.20
uritemplate==3.0.0
xlrd==1.0.0
